refactor: remove play_game method call from init
Removed the play_game method call from the 'Initialize' method
definition in the Board class. Because the method is now being called
from the start_game method in the main.rb file, which allows the user
to start a game without having to 'require' the board.rb file while
using the 'irb' or 'pry' command, since an instance of the Board class
is now automatically being created.